Biobjective Binary Programs

Biobjective Assignment Problems (BOAP)

MOOLib.read_boap_hadi_instance — Method.

read_boap_hadi_instance(filename::AbstractString)

Returns a BOBPInstance of a Bi Objective Assignment Problem ( Hadi Instance ) corresponding to filename

source

MOOLib.read_boap_hadi_non_dom_pts — Method.

read_boap_hadi_non_dom_pts(filename::AbstractString)

Returns True Non Dominated Points of a Bi Objective Assignment Problem ( Hadi Instance )

source

MOOLib.read_boap_hadi — Method.

read_boap_hadi(num::Int64)

Returns a BOBPInstance of a Bi Objective Assignment Problem ( Hadi Instance ) and its corresponding True Non Dominated Points

source

Biobjective Knapsack Problems (BOKP)

MOOLib.read_bokp_hadi_instance — Method.

read_bokp_hadi_instance(filename::AbstractString)

Returns a BOBPInstance of a Bi Objective Knapsack Problem ( Hadi Instance ) corresponding to filename

source

MOOLib.read_bokp_hadi_non_dom_pts — Method.

read_bokp_hadi_non_dom_pts(filename::AbstractString)

Returns True Non Dominated Points of a Bi Objective Knapsack Problem ( Hadi Instance )

source

MOOLib.read_bokp_hadi — Method.

read_bokp_hadi(num::Int64)

Returns a BOBPInstance of a Bi Objective Knapsack Problem ( Hadi Instance ) and its corresponding True Non Dominated Points

source

read_bokp_xavier_instance1(filename::AbstractString)

read_bokp_xavier_instance2(filename::AbstractString)

MOOLib.read_bokp_xavier_non_dom_pts — Method.

read_bokp_xavier_non_dom_pts(filename::AbstractString)

Returns True Non Dominated Points of a Bi Objective Knapsack Problem ( Xavier Instance )

source

MOOLib.read_bokp_xavier1 — Method.

read_bokp_xavier1(name::String)

Returns a BOBPInstance of a Bi Objective Knapsack Problem ( Xavier Class 1 Instance ) and its corresponding True Non Dominated Points

source

MOOLib.read_bokp_xavier2 — Method.

read_bokp_xavier2(name::String)

Returns a BOBPInstance of a Bi Objective Knapsack Problem ( Xavier Class 2 Instance ) and its corresponding True Non Dominated Points

source

Biobjective Set Covering Problems (BOSCP)

MOOLib.read_boscp_xavier_instance — Method.

read_boscp_xavier_instance(n::Int64, m::Int64, instance)

Returns a BOBPInstance of a Bi Objective Set Covering Problem ( Xavier Instance ) corresponding to n, m and instance.

source

MOOLib.read_boscp_xavier_non_dom_pts — Method.

read_boscp_xavier_non_dom_pts(filename::AbstractString)

Returns True Non Dominated Points of a Bi Objective Set Covering Problem ( Xavier Instance )

source

MOOLib.read_boscp_xavier — Method.

read_boscp_xavier(n::Int64, m::Int64, instance::Int64)

Returns a BOBPInstance of a Bi Objective Set Covering Problem ( Xavier Instance ) and its corresponding True Non Dominated Points.

source

Biobjective Set Packing Problems (BOSPP)

MOOLib.read_bospp_xavier_instance — Method.

read_bospp_xavier_instance(filename::AbstractString)

Returns a BOBPInstance of a Bi Objective Set Packing Problem ( Xavier Instance ) corresponding to filename.

source

MOOLib.read_bospp_xavier_non_dom_pts — Method.

read_bospp_xavier_non_dom_pts(filename::AbstractString)

Returns True Non Dominated Points of a Bi Objective Set Packing Problem ( Xavier Instance )

source

MOOLib.read_bospp_xavier — Method.

read_bospp_xavier(instance_id::String)

Returns a BOBPInstance of a Bi Objective Set Packing Problem ( Xavier Instance ) and its corresponding True Non Dominated Points.

source

Biobjective Multi Dimensional Knapsack Problems (BOKP)

For this set of instances, the true non dominated frontier is not available.

MOOLib.read_bokp_aritra_instance — Method.

read_bokp_aritra_instance(n::Int64, m::Int64, instance::Int64)

Returns a BOBPInstance of a Bi Objective Knapsack Problem ( Aritra Instances ) corresponding to n, m and instance.

source